backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.4.0
python-dotenv==1.0.1
werkzeug==3.0.3
requests==2.32.3
easyocr==1.7.1
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
easyocr==1.7.1
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
